+++ bucket-api/src/config/ormConfig.ts
@@ -20,7 +20,7 @@ export const ORMConfig: ConnectionOptions = {
   migrations: ["./src/thing/migration/**/*.ts"],
   subscribers: ["./src/thing/subscriber/**/*.ts"],
   // cli: {
-  //     entitiesDir: "./thing/entities",
+  //     entitiesDir: "./thing",
   //     migrationsDir: "./thing/migration",
   //     subscribersDir: "./thing/subscriber"
   // }