+++ bucket-api/src/http/HttpAPI.ts
@@ -59,7 +59,7 @@ export class HttpAPI {
      * @apiGroup DPi
      * @apiDescription Health status of the DPi Generator (available or not available)
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      **/
     this.app.get(
       config.http.baseUrl + "/things/types/dpi/health",
@@ -77,7 +77,7 @@ export class HttpAPI {
      * @apiGroup Property
      * @apiDescription List all properties accessible for the authenticated person.
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      *
      * @apiHeader {String} Authorization TOKEN ID
      *