+++ bucket-api/src/thing/property/PropertyRouter.ts
@@ -48,7 +48,7 @@ export class PropertyRouter {
      * @apiGroup Property
      * @apiDescription Get Properties of a Thing.
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      *
      * @apiHeader {String} Authorization TOKEN ID
      *
@@ -68,7 +68,7 @@ export class PropertyRouter {
      * @apiGroup Property
      * @apiDescription Get one Property.
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      *
      * @apiHeader {String} Authorization TOKEN ID
      *
@@ -94,7 +94,7 @@ export class PropertyRouter {
      * @apiGroup Property
      * @apiDescription Get the media associated to a dimension's timestamp.
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      *
      * @apiHeader {String} Authorization TOKEN ID
      *
@@ -138,7 +138,7 @@ export class PropertyRouter {
      * @apiGroup Property
      * @apiDescription Create a Property.
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      * @apiParam {String} thingId Id of the Thing to which we add the Property.
      *
      * @apiParam (Body) {Property} property Property to create as JSON.
@@ -168,7 +168,7 @@ export class PropertyRouter {
      * @apiGroup Property
      * @apiDescription Edit one Property to change its name or description.
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      *
      * @apiHeader {String} Authorization TOKEN ID
      *
@@ -196,7 +196,7 @@ export class PropertyRouter {
      * @apiGroup Property
      * @apiDescription Update values of a Property.
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      *
      * @apiHeader {String} Authorization TOKEN ID
      *
@@ -224,7 +224,7 @@ export class PropertyRouter {
      * @apiGroup Property
      * @apiDescription Delete one Property.
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      *
      * @apiHeader {String} Authorization TOKEN ID
      *
@@ -238,6 +238,27 @@ export class PropertyRouter {
         this.policyController.checkPolicy("delete"),
       ],
       this.controller.deleteOneProperty.bind(this.controller)
+    );
+
+    /**
+     * @api {delete} /things/:thingId/properties/:propertyId/timestamps Delete
+     * @apiGroup Property
+     * @apiDescription Delete some data points of a property.
+     *
+     * @apiVersion 0.1.5
+     *
+     * @apiHeader {String} Authorization TOKEN ID
+     *
+     * @apiParam {String} thingId Id of the Thing containing the Property from where to delete data points.
+     * @apiParam {String} propertyId Id of the Property from where to delete data points.
+     **/
+    this.router.delete(
+      "/:propertyId/timestamps",
+      [
+        this.authController.authenticate(["dcd:properties"]),
+        this.policyController.checkPolicy("delete"),
+      ],
+      this.controller.deleteDataPoints.bind(this.controller)
     );
 
     this.router.get(
@@ -263,7 +284,7 @@ export class PropertyRouter {
      * @apiGroup Property
      * @apiDescription List consents granted for one Property. Only property owner can access this list.
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      *
      * @apiHeader {String} Authorization TOKEN ID
      *
@@ -284,7 +305,7 @@ export class PropertyRouter {
      * @apiGroup Property
      * @apiDescription Revoke a consent granted for one Property. Only property owner can access this list.
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      *
      * @apiHeader {String} Authorization TOKEN ID
      *
@@ -306,7 +327,7 @@ export class PropertyRouter {
      * @apiGroup Property
      * @apiDescription Grant a consent for one Property. Only property owner can access this list.
      *
-     * @apiVersion 0.1.4
+     * @apiVersion 0.1.5
      *
      * @apiParam (Body) {Consent} consent Consent to grant as JSON.
      * @apiParamExample {json} consent: