+++ CHANGE_LOG.md
@@ -12,6 +12,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added rate limte on HTTP requests to mitigate DOS
 - Support video streaming WebRTC (commented out, work in progress)
 
+## [0.1.4] - 2022-03-15
+
+### Added
+
+- Delete data with property
+- Delete specific timestamp
+- PropertyService test
+
 ## [0.1.4] - 2022-02-15
 
 - small fixes